Move internal/x/cfginit to internal/cfginit
