+++ internal/x/exec/runner.go
@@ -39,8 +39,8 @@ import (
 
 	"github.com/golang/protobuf/jsonpb"
 	"github.com/golang/protobuf/protoc-gen-go/descriptor"
+	"github.com/uber/prototool/internal/cfginit"
 	"github.com/uber/prototool/internal/vars"
-	"github.com/uber/prototool/internal/x/cfginit"
 	"github.com/uber/prototool/internal/x/diff"
 	"github.com/uber/prototool/internal/x/extract"
 	"github.com/uber/prototool/internal/x/file"
